clippy

+++ src/bitmap/mod.rs
@@ -72,7 +72,7 @@ impl PdfiumBitmap {
     /// Returns the width of the image in the bitmap buffer backing this [`PdfiumBitmap`].
     #[inline]
     pub fn width(&self) -> i32 {
-        lib().FPDFBitmap_GetWidth(self) as i32
+        lib().FPDFBitmap_GetWidth(self)
     }
 
     /// Returns the height of the image in the bitmap buffer backing this [`PdfiumBitmap`].
